+++ back/src/alarms/alarms.service.ts
@@ -1,11 +1,6 @@
-import {
-  ForbiddenException,
-  Injectable,
-  NotFoundException,
-} from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { Agenda } from 'agenda/es';
-import { FcmService } from 'src/fcm/fcm.service';
+import { AgendaService } from 'src/agenda/agenda.service';
 import { PrismaMongoService } from 'src/prisma/prisma-mongo.service';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { DeleteAlarmsDto, SetAlarmDto } from './dto';
@@ -16,73 +11,46 @@ export class AlarmsService {
     private readonly prismaMongo: PrismaMongoService,
     private readonly prisma: PrismaService,
     private readonly configService: ConfigService,
-    private readonly fcmService: FcmService,
+    private readonly agnedaService: AgendaService,
   ) {}
 
-  async setAgenda(id: string, pillBookmarkId: string): Promise<Agenda> {
-    const agenda = new Agenda({
-      db: {
-        address: this.configService.get('DATABASE_URL_MONGO'),
-        collection: 'pillAlarms',
-      },
-      name: id + '-' + pillBookmarkId + 'set',
-    });
-    return agenda;
-  }
-
+  /** 현재 로그인한 유저의 북마크한 알약 알림 설정 */
   async setAlarms(id: string, setAlarmDto: SetAlarmDto) {
     const {
       vip,
       hour,
       minute,
+      repeatTime,
       pillName,
       userName,
-      repeatTime,
       deviceToken,
       pillBookmarkId,
     } = setAlarmDto;
 
     await this.saveDevicetoken(deviceToken, id, pillBookmarkId);
-    const agenda = await this.setAgenda(id, pillBookmarkId);
-
     try {
-      agenda.define(id + '-' + pillBookmarkId, async () => {
-        await this.fcmService.sendPushAlarm(deviceToken, userName, pillName);
-        const time = await this.getCurrTime();
-
-        await this.prismaMongo.reminder.create({
-          data: {
-            user_id: id,
-            user_name: userName,
-            pill_name: pillName,
-            time,
-          },
-        });
-
-        if (repeatTime !== 0) {
-          await this.setRepeatAlarm(repeatTime, agenda, id, pillBookmarkId);
-        }
-
-        if (vip.length === 8) {
-          await agenda.cancel({ name: id + '-' + pillBookmarkId });
-        }
-      });
+      await this.agnedaService.defineEveryAgenda(
+        id,
+        pillBookmarkId,
+        deviceToken,
+        userName,
+        pillName,
+        repeatTime,
+        vip,
+      );
 
       if (vip.length === 0) {
         vip.push(0, 1, 2, 3, 4, 5, 6, 7);
       }
 
-      (async function () {
-        await agenda.start();
-        await agenda.every(
-          `${minute} ${hour} * * ${vip.join(',')}`,
-          id + '-' + pillBookmarkId,
-          {
-            repeatTime,
-            timezone: 'Asia/Seoul',
-          },
-        );
-      })();
+      await this.agnedaService.setEveryAgenda(
+        minute,
+        hour,
+        vip,
+        repeatTime,
+        id,
+        pillBookmarkId,
+      );
 
       await this.setAlarmMark(true, pillBookmarkId);
     } catch (error) {
@@ -113,27 +81,9 @@ export class AlarmsService {
     }
   }
 
-  async getCurrTime() {
-    const fullTime = new Date(Date.now());
-    const year = fullTime.getFullYear().toString();
-    const month = (fullTime.getMonth() + 1).toString();
-    const date = fullTime.getDate().toString();
-    const hour = ('0' + fullTime.getHours()).slice(-2);
-    const minute = ('0' + fullTime.getMinutes()).slice(-2);
-
-    const time = year + '.' + month + '.' + date + ' ' + hour + ':' + minute;
-    return time;
-  }
-
   async cancelAlarm(id: string, pillBookmarkId: string) {
-    const agenda = await this.setAgenda(id, pillBookmarkId);
-
-    try {
-      await agenda.cancel({ name: id + '-' + pillBookmarkId });
-      await agenda.cancel({ name: id + '-' + pillBookmarkId + ':repeat' });
-    } catch (error) {
-      throw new NotFoundException('알림을 삭제하지 못했습니다.');
-    }
+    await this.agnedaService.deleteAgenda(id, pillBookmarkId);
+    await this.setAlarmMark(false, pillBookmarkId);
   }
 
   async setAlarmMark(check: boolean, pillBookmarkId: string) {
@@ -144,23 +94,6 @@ export class AlarmsService {
       });
     } catch (error) {
       throw new ForbiddenException('알림을 체크하지 못했습니다.');
-    }
-  }
-
-  async setRepeatAlarm(
-    repeatTime: number,
-    agenda: Agenda,
-    id: string,
-    pillBookmarkId: string,
-  ) {
-    try {
-      (async function () {
-        const job = agenda.create(id + '-' + pillBookmarkId + ':repeat');
-        await agenda.start();
-        await job.schedule(`in ${repeatTime} minutes`).save();
-      })();
-    } catch (error) {
-      throw new ForbiddenException('반복 시간을 설정하지 못했습니다.');
     }
   }
 
@@ -200,40 +133,8 @@ export class AlarmsService {
   }
 
   async getSetAlarm(id: string, pillBookmarkId: string) {
-    const agenda = await this.setAgenda(id, pillBookmarkId);
     const pillName = await this.getPillName(pillBookmarkId);
-    try {
-      const result = (async function () {
-        await agenda.start();
-        const job = await agenda.jobs({ name: id + '-' + pillBookmarkId });
-        if (job.length === 0) {
-          return {
-            minute: 0,
-            hour: 0,
-            vip: [],
-            repeatTime: 0,
-            pillName,
-          };
-        } else {
-          const alarm = job.pop().attrs;
-
-          const arr = alarm.repeatInterval.split(' ');
-          const repeat = alarm.data.repeatTime;
-
-          return {
-            minute: Number(arr[0]),
-            hour: Number(arr[1]),
-            vip: arr[4].split(',').map((v) => Number(v)),
-            repeatTime: repeat,
-            pillName,
-          };
-        }
-      })();
-
-      return result;
-    } catch (error) {
-      throw new ForbiddenException('알림을 조회하지 못했습니다.');
-    }
+    return await this.agnedaService.getAgenda(id, pillBookmarkId, pillName);
   }
 
   async getPillName(id: string) {
