+++ back/src/users/users.module.ts
@@ -1,4 +1,6 @@
 import { MiddlewareConsumer, Module } from '@nestjs/common';
+import { AgendaModule } from 'src/agenda/agenda.module';
+import { AgendaService } from 'src/agenda/agenda.service';
 import { AlarmsModule } from 'src/alarms/alarms.module';
 import { AlarmsService } from 'src/alarms/alarms.service';
 import { LoggerMiddleware } from 'src/common/middlewares/LoggerMiddleware';
@@ -8,10 +10,10 @@ import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
 
 @Module({
-  imports: [GcsModule, AlarmsModule, RedisModule],
-  providers: [UsersService, AlarmsService],
+  imports: [GcsModule, AlarmsModule, RedisModule, AgendaModule],
+  providers: [UsersService, AlarmsService, AgendaService],
   controllers: [UsersController],
-  exports: [GcsModule, AlarmsModule, RedisModule],
+  exports: [GcsModule, AlarmsModule, RedisModule, AgendaModule],
 })
 export class UsersModule {
   configure(consumer: MiddlewareConsumer) {
